Made form into compoment
